backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.0.0
python-multipart==0.0.12
email-validator==2.0.0
pytest==8.0.0
httpx==0.27.0
